+++ integrations/azure-devops/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.4.1 (2025-09-16)
+
+
+### Improvements
+
+- Added support for test-run kind in Azure DevOps integration
+- Added enrichment to include test results
+
+
 ## 0.4.0 (2025-09-11)